+++ packages/electron/desktop/src/main.tsx
@@ -1,8 +1,0 @@
-import { render } from "preact";
-import { App } from "./App";
-import { applyPlatformClass } from "./platform";
-import "./index.css";
-
-applyPlatformClass();
-
-render(<App />, document.getElementById("app")!);